+++ modules/shadertools/test/index.ts
@@ -2,26 +2,28 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) vis.gl contributors
 
+// SHADER MODULE API
+
+import './lib/shader-module/shader-module.spec';
+import './lib/shader-module/shader-module-dependencies.spec';
+
 // utils
 import './lib/glsl-utils/get-shader-info.spec';
 import './lib/glsl-utils/shader-utils.spec';
 
-// shader transpilation
-import './lib/preprocessor/preprocessor.spec';
+// WGSL parsing
+import './lib/wgsl/get-shader-layout-wgsl.spec';
 
 // shader transpilation
+import './lib/preprocessor/preprocessor.spec';
 import './lib/shader-transpiler/transpile-shader.spec';
 
 // shader generation
 import './lib/generator/generate-shader.spec';
 
-// WGSL parsing
-import './lib/wgsl/get-shader-layout-wgsl.spec';
 
-// SHADER MODULE API
+// SHADER ASSEMBLY
 
-import './lib/shader-module/shader-module.spec';
-import './lib/shader-assembly/resolve-modules.spec';
 import './lib/shader-assembly/inject-shader.spec';
 import './lib/shader-assembly/assemble-shaders.spec';
 import './lib/shader-assembler.spec';
